Correcciones

+++ CleanDataApi/src/main/java/utl/org/ldsm504/sakura/CleanDataApi/servicio/CamionServicioImp.java
@@ -1,18 +1,25 @@
 package utl.org.ldsm504.sakura.CleanDataApi.servicio;
 
+import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 import utl.org.ldsm504.sakura.CleanDataApi.modelo.Camion;
 import utl.org.ldsm504.sakura.CleanDataApi.modelo.EstadoCamion;
+import utl.org.ldsm504.sakura.CleanDataApi.repositorio.CamionGPSRepositorio;
 import utl.org.ldsm504.sakura.CleanDataApi.repositorio.CamionRepositorio;
+import utl.org.ldsm504.sakura.CleanDataApi.repositorio.ViajeRepositorio;
 
 import java.util.List;
 @Service
 public class CamionServicioImp implements CamionServicio{
 
     private final CamionRepositorio camionRepositorio;
+    private final ViajeRepositorio viajeRepositorio;
+    private final CamionGPSRepositorio camionGPSRepositorio;
 
-    public CamionServicioImp(CamionRepositorio camionRepositorio) {
+    public CamionServicioImp(CamionRepositorio camionRepositorio, ViajeRepositorio viajeRepositorio, CamionGPSRepositorio camionGPSRepositorio) {
         this.camionRepositorio = camionRepositorio;
+        this.viajeRepositorio = viajeRepositorio;
+        this.camionGPSRepositorio = camionGPSRepositorio;
     }
 
     @Override
@@ -31,6 +38,7 @@ public class CamionServicioImp implements CamionServicio{
         return camionRepositorio.findAll();
     }
 
+    @Transactional
     @Override
     public Camion actualizarCamion(Camion camion) {
         if (camion.getIdCamion() == null) {
@@ -53,6 +61,12 @@ public class CamionServicioImp implements CamionServicio{
         if (!camionRepositorio.existsById(id)) {
             throw new RuntimeException("No existe camión con id " + id);
         }
+        if (viajeRepositorio.existsByCamionIdCamion(id)) {
+            throw new RuntimeException("No se puede eliminar el camión porque está asociado a uno o más viajes");
+        }
+        if (camionGPSRepositorio.existsByCamionIdCamion(id)) {
+            throw new RuntimeException("No se puede eliminar el camión porque está asociado a un dispositivo GPS");
+        }
         camionRepositorio.deleteById(id);
     }
 
@@ -61,6 +75,7 @@ public class CamionServicioImp implements CamionServicio{
         return camionRepositorio.findByEstado(estadoCamion);
     }
 
+    @Transactional
     @Override
     public Camion actualizarCamionPorId(Integer id, Camion datos) {
         Camion existente = obtenerCamionPorId(id);
